refactor(wshandler): split wshandler.py into per-concern sub-package (#965)
Split the 3175-line wshandler.py monolith into a wshandler/ package
with per-concern submodules, following the same pattern used for
model/ in #993.

Submodules: _constants, safe_websocket, session, controllers,
connection, agent_mention, dispatch, helpers. The __init__.py
re-exports all names so existing imports are unchanged.
