+++ src/backend/klangk_backend/wshandler/agent_mention.py
@@ -1,0 +1,147 @@
+"""Agent-mention detection and routing."""
+
+import asyncio
+import logging
+import re
+
+from .. import agent, model
+from ._constants import (
+    _agent_conversations as _agent_conversations,
+    _agent_tasks as _agent_tasks,
+    _cancel_agent_task as _cancel_agent_task,
+    _drop_agent_task_if_current as _drop_agent_task_if_current,
+)
+from .session import state
+
+logger = logging.getLogger(__name__)
+
+
+# _ANY_MENTION_RE matches any @mention; the agent-specific regex is
+# compiled fresh per call (see _get_agent_mention_re).
+_ANY_MENTION_RE = re.compile(r"(?:^|(?<=\s))@\S+")
+
+
+def _get_agent_mention_re(handle: str) -> re.Pattern:
+    """Return the compiled agent mention regex for *handle*.
+
+    The agent's handle can change at runtime (DB update), so this is
+    compiled fresh on every call rather than cached. Regex compilation
+    is cheap, and caching would return a stale pattern after a rename.
+    """
+    return re.compile(
+        r"(?:^|(?<=\s))@" + re.escape(handle) + r"(?:\s|$)",
+        re.IGNORECASE,
+    )
+
+
+async def _mentions_agent(text: str) -> bool:
+    """Return True if the message text mentions the agent."""
+    handle = await model.agent_handle()
+    return bool(_get_agent_mention_re(handle).search(text))
+
+
+async def _addresses_other_user(text: str) -> bool:
+    """Return True if the message is directed at someone else.
+
+    A message that *starts* with ``@someone`` (not the agent) is
+    considered addressed to that person, breaking the follow-up
+    conversation with the agent.
+    """
+    m = _ANY_MENTION_RE.match(text.lstrip())
+    if not m:
+        return False
+    mention = m.group().lstrip("@").lower()
+    handle = await model.agent_handle()
+    return mention != handle.lower()
+
+
+async def _handle_agent_mention(
+    workspace_id: str, container_id: str, user_text: str
+) -> None:
+    """Handle an @agent mention by sending the prompt to Pi RPC."""
+
+    agent_handle = await model.agent_handle()
+    agent_re = _get_agent_mention_re(agent_handle)
+    prompt = agent_re.sub("", user_text).strip()
+    if not prompt:
+        prompt = "Hello!"
+
+    # Include messages from OTHER users since the agent's last response
+    # as context.  The current user's message is already the prompt;
+    # we only need to show interjections from other participants that
+    # Pi hasn't seen (since Pi's multi-turn history only has the
+    # conversation between the mentioning user and itself).
+    recent = await model.get_chat_messages(workspace_id, limit=50)
+    chronological = recent
+    last_agent_idx = -1
+    for i, m in enumerate(chronological):
+        if m.get("message_type", 0) == model.MSG_AGENT:
+            last_agent_idx = i
+    # Messages from other users (not the current prompt sender)
+    other_msgs = [
+        m
+        for m in chronological[last_agent_idx + 1 :]
+        if m.get("message_type", 0) == model.MSG_USER
+        and m.get("message", "").strip() != user_text.strip()
+    ]
+    if other_msgs:
+        context_lines = [
+            f"{m.get('user_email', 'unknown')}: {m.get('message', '')}"
+            for m in other_msgs
+        ]
+        context = "\n".join(context_lines)
+        prompt = f"[Other participants said:\n{context}]\n\n{prompt}"
+
+    agent_email = await model.agent_email()
+
+    # Notify clients the agent is thinking
+    session = state.get_session(workspace_id)
+    if session:
+        session.broadcast(
+            {
+                "type": "agent_thinking",
+                "thinking": True,
+                "name": agent_handle,
+            }
+        )
+
+    try:
+        pi = await agent.get_session(workspace_id)
+        response_text = await pi.send_prompt(prompt)
+    except asyncio.CancelledError:  # pragma: no cover
+        response_text = "Stopped."
+    except agent.AgentProcessDied:
+        logger.warning("Agent process died for workspace %s", workspace_id)
+        # Broadcast an ephemeral (non-persisted) system message — agent
+        # presence transitions are container-lifecycle events and must
+        # not linger in chat history (no symmetric persisted "connected").
+        sys_msg = agent._ephemeral_system_message(
+            workspace_id,
+            agent_email,
+            agent_handle,
+            f"{agent_handle} has disconnected",
+        )
+        session = state.get_session(workspace_id)
+        if session:
+            session.broadcast({"type": "agent_thinking", "thinking": False})
+            session.broadcast({"type": "chat_message", **sys_msg})
+        _drop_agent_task_if_current(workspace_id)
+        return
+    except Exception:
+        logger.exception("Agent error for workspace %s", workspace_id)
+        response_text = (
+            "Sorry, I encountered an error processing your request."
+        )
+
+    agent_msg = await model.add_chat_message(
+        workspace_id,
+        model.AGENT_USER_ID,
+        agent_email,
+        response_text,
+        message_type=model.MSG_AGENT,
+    )
+    session = state.get_session(workspace_id)
+    if session:
+        session.broadcast({"type": "agent_thinking", "thinking": False})
+        session.broadcast({"type": "chat_message", **agent_msg})
+    _drop_agent_task_if_current(workspace_id)
